+++ bindings/nodejs/tests/client/utilityMethods.spec.ts
@@ -97,12 +97,12 @@ describe('Client utility methods', () => {
         const aliasId =
             '0xcf077d276686ba64c0404b9eb2d15556782113c5a1985f262b70f9964d3bbd7f';
         const serialNumber = 0;
-        const tokenSchemeKind = 0;
+        const tokenSchemeType = 0;
 
         const foundryId = Utils.computeFoundryId(
             aliasId,
             serialNumber,
-            tokenSchemeKind,
+            tokenSchemeType,
         );
 
         expect(foundryId).toBe(