+++ sdk/src/types/block/error.rs
@@ -80,7 +80,7 @@ pub enum Error {
     InvalidTagFeatureLength(<TagFeatureLength as TryFrom<usize>>::Error),
     InvalidTagLength(<TagLength as TryFrom<usize>>::Error),
     InvalidTailTransactionHash,
-    InvalidTokenSchemeKind(u8),
+    InvalidTokenSchemeType(u8),
     InvalidTransactionAmountSum(u128),
     InvalidTransactionNativeTokensCount(u16),
     InvalidTreasuryOutputAmount(u64),
@@ -233,7 +233,7 @@ impl fmt::Display for Error {
                 write!(f, "invalid tag length {length}")
             }
             Self::InvalidTailTransactionHash => write!(f, "invalid tail transaction hash"),
-            Self::InvalidTokenSchemeKind(k) => write!(f, "invalid token scheme kind {k}"),
+            Self::InvalidTokenSchemeType(k) => write!(f, "invalid token scheme type {k}"),
             Self::InvalidTransactionAmountSum(value) => write!(f, "invalid transaction amount sum: {value}"),
             Self::InvalidTransactionNativeTokensCount(count) => {
                 write!(f, "invalid transaction native tokens count: {count}")