+++ sdk/tests/types/foundry_id.rs
@@ -19,7 +19,7 @@ fn getters() {
 
     assert_eq!(foundry_id.alias_address(), alias_address);
     assert_eq!(foundry_id.serial_number(), serial_number);
-    assert_eq!(foundry_id.token_scheme_kind(), token_scheme.kind());
+    assert_eq!(foundry_id.token_scheme_type(), token_scheme.kind());
     assert_eq!(
         foundry_id,
         FoundryId::from_str("0x0852fdfc072182654f163f5f0f9a621d729566c74d10037c4d7bbb0407d1e2c6492a00000000").unwrap()